Fix broken Android LocalSocket implementation causing deadlock
When in a blocking read() call, closing the socket won't release
or make the read() call throw an exception. This will cause dead
locks if the socket has to change and the previous socket isn't
written to anymore. The workaround is to send a SHUTDOWN message
prior to shutting down the actual socket to make sure the previous
read loop gets broken.